[1-2 예외처리] 김민정 제출합니다. (#106)
* fix: theme 추가로 인한 링크 추가 및 기존 링크 수정

* feat: theme 추가로 인한 html 및 js 파일 생성

* feat: 실습 - 페이지 설정

* feat: 1-1 예약관리 : 추가

* feat: 1-1 예약관리 : 조회

* feat: 1-1 예약관리 : 삭제

* chore: 1-2 데이터 베이스 적용 : sql 추가

* feat : 1-2 데이터 베이스 적용 : 추가

* feat : 1-2 데이터 베이스 적용 : 조회

* feat : 1-2 데이터 베이스 적용 : 삭제

* chore : 1-3 시간 관리 기능 추가 : sql update

* chore : 1-3 시간 관리 기능 추가 : 패키지 변경

* feat : 1-3 시간 관리 기능 추가 : time 페이지 mapping

* feat : 1-3 시간 관리 기능 추가 : time 추가

* feat : 1-3 시간 관리 기능 추가 : time 조회

* feat : 1-3 시간 관리 기능 추가 : time 삭제

* modify : 1-3 시간 관리 기능 추가 : reservation 생성 시 timeId를 받도록 변경

* modify : 1-3 시간 관리 기능 추가 : reservation 조회 시 innerJoin을 사용한 결과를 반환하도록 수정

* fix : 1-3 시간 관리 기능 추가 : 요구사항에 맞게 생성 로직 수정

* fix : 1-3 시간 관리 기능 추가 : 프런트 코드에 맞게 reservation 생성 로직 변경

* fix : ReservationTime 테스트 코드 추가, 테스트 코드 깨지던 오류 수정

* refactor : @RequestMapping 적용하여 공통 url 제거

* refactor : 쿼리 multi line String 적용

* fix : reservation.html에 reservation-new.js 대신 reservation.js를 포함하도록 변경

* refactor : restController와 controller 분리

* refactor : 불필요한 생성자 삭제

* feat : 2-1 theme 추가

* feat : 2-1 theme 조회

* feat : 2-1 theme 삭제

* feat : theme - read, delete test

* feat : 2-1 변경된 요구사항에 맞게 reservation 수정

* feat : 시간 생성 시 유효하지 않은 값이 입력되었을 때 예외 처리

* feat : 시간 생성 시 유효하지 않은 값이 입력되었을 때 예외 처리 js alert

* feat : 예약 생성시 예외 처리 추가

* feat : 특정 시간에 대한 예약이 존재할 때 시간을 삭제하려는 경우 예외처리

* feat : 과거 시간에 예약을 생성하려는 경우 예외처리

* feat : 동일한 예약이 존재하는 경우 예외처리

* modify : test 클래스 public 제거

---------

+++ src/main/java/roomescape/service/ReservationTimeAdminService.java
@@ -1,29 +1,55 @@
 package roomescape.service;
 
 import org.springframework.stereotype.Service;
+import roomescape.exception.ReservationExistException;
+import roomescape.exception.ReservationTimeInvalidException;
 import roomescape.model.ReservationTime;
+import roomescape.model.ReservationTimeCreateRequestDto;
+import roomescape.respository.ReservationDAO;
 import roomescape.respository.ReservationTimeDAO;
 
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 @Service
 public class ReservationTimeAdminService {
     private final ReservationTimeDAO reservationTimeDao;
+    private final ReservationDAO reservationDAO;
 
-    public ReservationTimeAdminService(ReservationTimeDAO reservationTimeDao) {
+    public ReservationTimeAdminService(ReservationTimeDAO reservationTimeDao, ReservationDAO reservationDAO) {
         this.reservationTimeDao = reservationTimeDao;
+        this.reservationDAO = reservationDAO;
     }
 
-    public ReservationTime createReservationTime(ReservationTime time) {
+    public ReservationTime createReservationTime(ReservationTimeCreateRequestDto time) {
+        validateReservationTimeCreation(time);
         return reservationTimeDao.insertReservationTime(time.getStartAt());
+    }
+
+    private void validateReservationTimeCreation(ReservationTimeCreateRequestDto time) {
+        String startAt = time.getStartAt();
+        if (isInValidTime(startAt)) {
+            throw new ReservationTimeInvalidException("24시간제 형식에 맞는 시간을 입력해주세요.");
+        }
+    }
+
+    private boolean isInValidTime(String time) {
+        final String TIME_PATTERN = "^(?:[01]\\d|2[0-3]):[0-5]\\d$";
+
+        Pattern pattern = Pattern.compile(TIME_PATTERN);
+        Matcher matcher = pattern.matcher(time);
+        return !matcher.matches();
     }
 
     public List<ReservationTime> getReservationTime() {
         return reservationTimeDao.readReservationTime();
-
     }
 
     public void deleteReservationTime(Long id) {
+        if (reservationDAO.isReservationExistByTimeId(id)) {
+            throw new ReservationExistException("예약이 존재하는 시간은 삭제할 수 없습니다.");
+        }
         reservationTimeDao.deleteReservationTime(id);
     }
 }
